+++ ra4m1-ctpac/src/osm.rs
@@ -1,11 +1,10 @@
-#[doc = "Independent Watchdog Timer"]
 #[derive(Copy, Clone, Eq, PartialEq)]
-pub struct Iwdt {
+pub struct Osm {
     ptr: *mut u8,
 }
-unsafe impl Send for Iwdt {}
-unsafe impl Sync for Iwdt {}
-impl Iwdt {
+unsafe impl Send for Osm {}
+unsafe impl Sync for Osm {}
+impl Osm {
     #[inline(always)]
     pub const unsafe fn from_ptr(ptr: *mut ()) -> Self {
         Self { ptr: ptr as _ }
@@ -14,15 +13,16 @@ impl Iwdt {
     pub const fn as_ptr(&self) -> *mut () {
         self.ptr as _
     }
-    #[doc = "IWDT Refresh Register"]
+    #[doc = "Option Function Select Register 0 (OFS0)"]
     #[inline(always)]
-    pub const fn iwdtrr(self) -> crate::common::Reg<u8, crate::common::RW> {
+    pub const fn ofs0(self) -> crate::common::Reg<regs::Ofs0, crate::common::R> {
         unsafe { crate::common::Reg::from_ptr(self.ptr.wrapping_add(0x0usize) as _) }
     }
-    #[doc = "IWDT Status Register"]
+    #[doc = "Option Function Select Register 1 (OFS1)"]
     #[inline(always)]
-    pub const fn iwdtsr(self) -> crate::common::Reg<regs::Iwdtsr, crate::common::RW> {
+    pub const fn ofs1(self) -> crate::common::Reg<regs::Ofs1, crate::common::R> {
         unsafe { crate::common::Reg::from_ptr(self.ptr.wrapping_add(0x04usize) as _) }
     }
 }
 pub mod regs;
+pub mod vals;