+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project will eventually adhere to [Semantic Versioning](https://semver.org/spec/v2.0.0.html) once the 1.0 milestone is reached.
 
 ## [0.2.3] - TBD
+### Added
+- iwdt: new driver
+- wdt: new driver
+- pac: `OSM` peripheral with `OFS0` and `OFS1` registers
 
 ## [0.2.2] - 2026-03-02
 ### Added